+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### dependabot: \#31 Bump ip from 1.1.8 to 1.1.9
 
+- added `dateIsFirstDayOfMonth`, `dateIsLeapYear`, `dateIsWeekend` and `getNextBusinessDate` utility functions
+
 ## [0.4.0] - 2024-01-29
 
 - added `CacheProvider<TEnum>` utility class